+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Master
 
+- Updates AuctionCountDownTimer to handle liveStartAt - lilyfromseattle
+
 ### 1.17.7
 
 - Fixes crash with missing artist images - ash